+ use PuzzleInstanceCustom + rename to types/map.ts
