+++ src/components/App/MapRenderer/index.tsx
@@ -2,12 +2,11 @@ import React from 'react';
 import { Map, Marker, ZoomControl } from 'pigeon-maps';
 import { maptiler } from 'pigeon-maps/providers';
 import mapRendererStyles from './MapRenderer.module.scss';
-import {MapMarker, MapRendererProps} from "../../../types/mapRenderer";
+import { MapMarker, MapRendererProps } from '../../../types/map';
 
 const MAPTILER_ACCESS_TOKEN = process.env.NEXT_PUBLIC_MAPTILER_ACCESS_TOKEN;
 
 const mapTilerProvider = maptiler(MAPTILER_ACCESS_TOKEN, 'outdoor');
-
 
 const MapRenderer = ({
   markers,
