+++ src/types/map.ts
@@ -1,19 +1,10 @@
-import {Dispatch, SetStateAction} from "react";
-import {CardProps} from "./cards";
+import { Dispatch, SetStateAction } from 'react';
 
 export type MapAnchor = [number, number];
 
 export type MapMarker = {
   anchor: MapAnchor;
   zoom: number;
-};
-
-export type PuzzleInstance = {
-  id: string;
-  longitude: number;
-  latitude: number;
-  address: string;
-  puzzle: CardProps;
 };
 
 export type MapRendererProps = {
